display: Fetch displays list from server and find by title substring.
If the given display starts with a '/' it is used as the path directly.
Otherwise we fetch the JSON spec, and search the titles for a substring
match for the given display name.

For example, if you run with a display of `my-target`, it does not
start with a '/' so the JSON is fetched, and scanned. The first entry
with a title containing `my-target` would match.

So "FOO-BARmy-targetBAZ" would match.